net: Move local address functions to a new file
