+++ src/init.cpp
@@ -39,6 +39,7 @@
 #include <kernel/caches.h>
 #include <kernel/context.h>
 #include <key.h>
+#include <local_addresses.h>
 #include <logging.h>
 #include <mapport.h>
 #include <net.h>
