+++ src/test/fuzz/net.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <chainparams.h>
+#include <local_addresses.h>
 #include <net.h>
 #include <net_permissions.h>
 #include <netaddress.h>
