+++ src/test/util/setup_common.cpp
@@ -16,6 +16,7 @@
 #include <init/common.h>
 #include <interfaces/chain.h>
 #include <kernel/mempool_entry.h>
+#include <local_addresses.h>
 #include <logging.h>
 #include <net.h>
 #include <net_processing.h>
